update repositories methods

+++ ecommerceweb/src/main/java/com/backend/ecommerceweb/repository/IUserLoginRepository.java
@@ -4,6 +4,6 @@ import com.backend.ecommerceweb.entities.UserLogin;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface IUserLoginRepository extends JpaRepository<UserLogin, Long> {
-    UserLogin findUserLoginByUserId(Long id);
+    UserLogin findUserLoginByUserId(int id);
 
 }
