+++ ecommerceweb/src/main/java/com/backend/ecommerceweb/security/UserPrincipal.java
@@ -7,7 +7,7 @@ import java.util.Collection;
 
 @Data
 public class UserPrincipal implements UserDetails {
-    private Long userId;
+    private int userId;
     private String username;
     private String password;
     private Collection authorities;
