+++ ecommerceweb/src/main/java/com/backend/ecommerceweb/services/IUserLoginService.java
@@ -4,5 +4,5 @@ import com.backend.ecommerceweb.entities.UserLogin;
 
 public interface IUserLoginService {
     void save(UserLogin userLogin);
-    UserLogin findUserLoginByUserId(Long id);
+    UserLogin findUserLoginByUserId(int id);
 }
